// src/codegen/DaoProposalSingle.types.ts
export type Duration = {
  height: number;
} | {
  time: number;
};
export type PreProposeInfo = {
  anyone_may_propose: {};
} | {
  module_may_propose: {
    info: ModuleInstantiateInfo;
  };
};
export type Admin = {
  address: {
    addr: string;
  };
} | {
  core_module: {};
};
export type Uint128 = string;
export type Binary = string;
export type Threshold = {
  absolute_percentage: {
    percentage: PercentageThreshold;
  };
} | {
  threshold_quorum: {
    quorum: PercentageThreshold;
    threshold: PercentageThreshold;
  };
} | {
  absolute_count: {
    threshold: Uint128;
  };
};
export type PercentageThreshold = {
  majority: {};
} | {
  percent: Decimal;
};
export type Decimal = string;
export interface InstantiateMsg {
  allow_revoting: boolean;
  close_proposal_on_execution_failure: boolean;
  delegation_module?: string | null;
  max_voting_period: Duration;
  min_voting_period?: Duration | null;
  only_members_execute: boolean;
  pre_propose_info: PreProposeInfo;
  threshold: Threshold;
  veto?: VetoConfig | null;
}
export interface ModuleInstantiateInfo {
  admin?: Admin | null;
  code_id: number;
  funds?: Coin[] | null;
  label: string;
  msg: Binary;
  salt?: Binary | null;
}
export interface Coin {
  amount: Uint128;
  denom: string;
}
export interface VetoConfig {
  early_execute: boolean;
  timelock_duration: Duration;
  veto_before_passed: boolean;
  vetoer: string;
}